backend-api/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
celery==5.3.4
redis==5.0.1
Pillow==10.1.0
django-oauth-toolkit==1.7.1
cryptography==41.0.8
argon2-cffi==23.1.0
pyjwt==2.8.0
requests==2.31.0
django-extensions==3.2.3
gunicorn==21.2.0
